Crate wdk_macros

A collection of macros that help make it easier to interact with [wdk-sys]’s direct bindings to the Windows Driver Kit (WDK).

Macros

call_unsafe_wdf_function_binding

A procedural macro that allows WDF functions to be called by name.